refactor(dropzone-image): extract hotkeys logic into hook
